+++ api/utils.py
@@ -303,10 +303,13 @@ class APIUtils():
             'text/html': 2,
             'application/pdf': 3,
             'application/html+edd': 4,
-            'application/epub+xml': 5, 'application/epub+zip': 5,
+            'application/x.html+edd': 4,
+            'application/epub+xml': 5,
+            'application/epub+zip': 5,
+            'application/html+catalog': 6
         }
 
-        return scores[link['mediaType']]
+        return scores.get(link['mediaType'], 7)
 
     @classmethod
     def formatRecord(cls, record, itemsByLink):