+++ CHANGELOG.md
@@ -1,12 +1,11 @@
 # CHANGELOG
 
-## unreleased -- v0.10.5
+## 2022-07-21 -- v0.10.5
 ### Added
 - New script to update past IntechOpen records with new HTML links
 - Exposed entire authors object to the edition details page
 - Two new scripts to update NYPL catalogs links in Record and Link table
 - Updated NYPL mapping and formats in APIUtils with new catalog link
-- Government document filter for advanced search page
 - Script to update past works in ES with the proper boolean value in their is_government_document field
 ### Fixed
 - Added improved error handling around unexpected `media_type` values